move code for latin alphabet score to generic.rs
